+++ src/main/java/cc/cassian/raspberry/config/ModConfig.java
@@ -55,6 +55,7 @@ public class ModConfig {
     public boolean unified_wrenches = false;
     public boolean saferKnightJump = true;
     public boolean raspberryCartEngine = true;
+    public boolean dungeons_mobs_revised_cobwebs;
 
     public static void load() {
         if (!Files.exists(configPath())) {