+++ src/main/java/cc/cassian/raspberry/mixin/RaspberryMixinPlugin.java
@@ -46,6 +46,9 @@ public class RaspberryMixinPlugin implements IMixinConfigPlugin {
         else if (checkMixin("create", mixinClassName)){
             return false;
         }
+        else if (checkMixin("dungeons_mobs", mixinClassName)){
+            return false;
+        }
         else if (checkMixin("dynamiccrosshair", mixinClassName)){
             return false;
         }